feat: add Cost and Usage Glue visual ETL job (#22)
Add a download the Cost and Usage Glue visual ETL job definition.  There
is currently no way to manage visual ETL jobs using Terraform, but this will
give us a backup of the job definition in case we need to revert a change.

A future PR will have a nightly sync job that checks for new or modified visual ETL
jobs and creates automated PRs with the changes.