feat: glue job sync workflow (#28)
Add a GitHub workflow that retrieves the AWS Glue jobs and
checks if there are any new or changed jobs.  If there are, a PR is
opened with the changes.

Once this is tested and working, it will be set to run on a schedule
so that we have a backup of the visual Glue jobs, which currently
cannot be managed using Terraform.