fix: support the optional count fields in the Template data
Update the ETL so that it can properly handle the `_count` fields which
may or may not exist in each day's data extract.  If a count field is not present
it will be set to 0 for that day's data.

Additionally, the data write has been updated to no longer update the Glue
schema when the data is written back to S3.  This will prevent schema changes
that would break Superset reports as new `_count` fields appear.