fix: perform incremental loads at the partition level
Update the incremental load logic so that it considers the data
retention period and loads the full partition(s) required to
refresh the historical notification data each day.

So for example, if the current date is March 10th and the retention
period is 7 days, only the March partition will be reloaded.  However
if the date is March 3rd with the same 7 day retention period then both
March and April's partition data would be reloaded.