fix: add OIDC release role for production workflows (#255)
Add a new OIDC role that can be assumed by GitHub workflows that
are triggered by a code release.

Normally our OIDC roles for applying changes expect a claim of the `main`
branch, but in the case of our release workflows, the claim will be a GitHub
tag prefixed with `v`  (e.g. v1.1.0).

This new OIDC release role is only created in Production as it is not needed
in Staging.