fix: increase the retention days for the `notification_history` load
Double the amount of days the incremental load looks back and reloads month partitions
for.  This is being done because we noticed that on May 8/9, new records on April 29/30 were in
the `notification_history` table, but missed as the retention period of 7 days would only
have loaded the May data.